requirements.txt
apify-client>=1.6.0
nltk>=3.8.1
markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
playwright>=1.39.0
tenacity>=8.2.3
python-dotenv>=1.0.0 
datetime
pymongo>=4.6.0
urllib3>=2.1.0
dataclasses>=0.6
typing>=3.7.4.3
selenium>=4.15.2
scrapy>=2.11.0
openai>=1.3.0
pytest>=7.4.3
pytest-mock>=3.12.0
scikit-learn>=1.3.2
numpy>=1.26.2
free-proxy==1.1.1
transformers>=4.36.0
datasets>=2.16.0
torch>=2.1.0
accelerate>=0.25.0
huggingface-hub>=0.20.0